chore(auth): move fixtures to auth module
