+++ src/auth/pipes/swap-password.pipe.spec.ts
@@ -1,7 +1,7 @@
 import type { ArgumentMetadata } from '@nestjs/common';
 
+import { login as credentials } from '@/auth/fixtures/credentials';
 import { SwapPasswordPipe } from '@/auth/pipes/swap-password.pipe';
-import { credentials } from '@/common/fixtures/credentials';
 
 const value = {
   image: 'https://thispersondoesnotexist.com/image',
