+++ src/auth/services/authentication.service.spec.ts
@@ -6,9 +6,9 @@ import type { Repository } from 'typeorm';
 import type { RegisterUser } from '@/auth/dto/register-user.dto';
 import type { UpdateUser } from '@/auth/dto/update-user.dto';
 import { User } from '@/auth/entities/user.entity';
+import { login as credentials } from '@/auth/fixtures/credentials';
 import type { JwtPayload } from '@/auth/interfaces/jwt-payload.interface';
 import { AuthenticationService } from '@/auth/services/authentication.service';
-import { credentials } from '@/common/fixtures/credentials';
 
 const user = User.fromPartial({
   email: 'john@doe.me',
